+++ initdb.d/init-db.sh
@@ -5,6 +5,7 @@ DIR="/docker-entrypoint-initdb.d"
 PSQL="psql -U postgres"
 ${PSQL} -c "create database r99c owner='postgres'"
 ${PSQL} r99c < ${DIR}/r99c-2022-03-29.dump
-${PSQL} -c "alter table r99c.users drop colum if exist sid;"
-${PSQL} -c "alter table r99c.users drop colum if exist name;"
+# error `if exit`
+${PSQL} -c r99c "alter table users drop colum sid;"
+${PSQL} -c r99c "alter table users drop colum name;"
 ${PSQL} -c "alter database r99c set default_transaction_read_only = on;"